Fix a possible integer overflow
When I tried to use UndefinedSanitizer (by
CXXFLAGS="-fsanitize=undefined"), I found the following error.

```
% sudo ./src/bpftrace -e 'BEGIN{curtask}'
/home/ubuntu/work/bpftrace_master/src/clang_parser.cpp:225:22: runtime error: shift exponent 64 is too large for 32-bit type 'int'
SUMMARY: UndefinedBehaviorSanitizer: undefined-behavior /home/ubuntu/work/bpftrace_master/src/clang_parser.cpp:225:22 in
/home/ubuntu/work/bpftrace_master/src/clang_parser.cpp:225:44: runtime error: signed integer overflow: -2147483648 - 1 cannot be represented in type 'int'
SUMMARY: UndefinedBehaviorSanitizer: undefined-behavior /home/ubuntu/work/bpftrace_master/src/clang_parser.cpp:225:44 in
```

It warned about integer overflows. This patch fixes this.

+++ src/struct.h
@@ -15,7 +15,7 @@ struct Bitfield
   // Then rshift the resulting value by `access_rshift` to get field value
   size_t access_rshift;
   // Then logical AND `mask` to mask out everything but this bitfield
-  size_t mask;
+  uint64_t mask;
 };
 
 struct Field
